Rename WebSocket handler: forwarding is now the default base
Make CopilotWebSocketHandler the concrete handler that forwards to the
upstream Copilot service by default, and introduce
CopilotWebSocketHandlerBase as the lower-level abstraction that does no
upstream forwarding. Previously the forwarding behavior lived in
ForwardingCopilotWebSocketHandler while the base CopilotWebSocketHandler
did not forward, which made it non-obvious that overriding the handler
required switching to the Forwarding* type to preserve passthrough.

Applies the rename across the four object-oriented SDKs (Node, .NET,
Java, Python) and updates the corresponding e2e tests. Go and Rust use
composition and are unaffected.
