+++ python/copilot/__init__.py
@@ -69,7 +69,7 @@ from .copilot_request_handler import (
     CopilotRequestHandler,
     CopilotWebSocketCloseStatus,
     CopilotWebSocketHandler,
-    ForwardingCopilotWebSocketHandler,
+    CopilotWebSocketHandlerBase,
     LlmInferenceHeaders,
     create_copilot_request_adapter,
 )
@@ -198,7 +198,7 @@ __all__ = [
     "CopilotRequestContext",
     "CopilotRequestHandler",
     "CopilotWebSocketCloseStatus",
-    "CopilotWebSocketHandler",
+    "CopilotWebSocketHandlerBase",
     "CreateSessionFsHandler",
     "ElicitationContext",
     "ElicitationHandler",
@@ -211,7 +211,7 @@ __all__ = [
     "ExitPlanModeRequest",
     "ExitPlanModeResult",
     "ExtensionInfo",
-    "ForwardingCopilotWebSocketHandler",
+    "CopilotWebSocketHandler",
     "GetAuthStatusResponse",
     "GetStatusResponse",
     "InfiniteSessionConfig",
